requirements.txt
torch==1.12.1
dgl==0.8.1
pandas==1.3.5
numpy==1.21.6
scikit-learn==1.0.2
scipy==1.7.3
networkx==2.6.3
pyyaml==6.0
matplotlib
tqdm
